+++ lib/dartactyl.dart
@@ -5,6 +5,6 @@ library dartactyl;
 
 export 'models.dart';
 export 'src/clients/application.dart';
-export 'src/clients/client.dart';
+export 'src/clients/client.dart' hide PteroClientVisibleDio;
 export 'src/clients/translation_client.dart';
 export 'src/websocket/server_websocket.dart';