+++ lib/dartactyl_visible_dio.dart
@@ -1,0 +1,4 @@
+library dartactyl;
+
+export 'dartactyl.dart';
+export 'src/clients/client.dart' show PteroClientVisibleDio;